Restructured interpreter. Added StrawberryAllocator for memory allocation and handling of visitor return values. StrawberryInterpreter for visitor overrides and handling script logic.
