+++ Strawberry/Cpp/StrawberryInterpreter.cpp
@@ -1,5 +1,5 @@
 //
-// Created by patrick on 3/10/23.
+// Created by Patrick on 4/6/2023.
 //
 
 #include "StrawberryInterpreter.h"
@@ -13,6 +13,7 @@
 #include "types/expressions/Bool.h"
 
 namespace antlrcpptest {
+
 /* ================================================ VISITOR OVERRIDES =============================================== */
     std::any StrawberryInterpreter::visitScript_(StrawberryParser::Script_Context *ctx) {
         this->scope_in();
@@ -23,6 +24,15 @@ namespace antlrcpptest {
         std::cout << "Default =============================================================" << std::endl;
         print_default();
         return 0;
+    }
+
+    std::any StrawberryInterpreter::visitAction_(StrawberryParser::Action_Context *ctx) {
+        std::cout << ">> " << ctx->getText() << std::endl;
+        auto result =  StrawberryParserBaseVisitor::visitAction_(ctx);
+        print_memory();
+        print_default("  --> ");
+        std::cout << std::endl << std::endl;
+        return result;
     }
 
     std::any StrawberryInterpreter::visitScope(StrawberryParser::ScopeContext *ctx) {
@@ -36,8 +46,8 @@ namespace antlrcpptest {
         return action_response;
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Loop Keywords */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Loop Keywords */
     std::any StrawberryInterpreter::visitReturnStat(StrawberryParser::ReturnStatContext *ctx) {
         return StrawberryParserBaseVisitor::visitReturnStat(ctx);
     }
@@ -58,8 +68,8 @@ namespace antlrcpptest {
         return StrawberryParserBaseVisitor::visitBreakStat(ctx);
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Control Flow */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Control Flow */
     std::any StrawberryInterpreter::visitLoopScope(StrawberryParser::LoopScopeContext *ctx) {
         return StrawberryParserBaseVisitor::visitLoopScope(ctx);
     }
@@ -128,11 +138,11 @@ namespace antlrcpptest {
         return StrawberryParserBaseVisitor::visitUntilLoop(ctx);
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Function Handlers */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Function Handlers */
     std::any StrawberryInterpreter::visitFnDeclaration(StrawberryParser::FnDeclarationContext *ctx) {
         auto fn = std::make_unique<FunctionHandle>(shared_from_this(), ctx);
-        this->functionLibrary->add(ctx->Id()->getText(),std::move(fn));
+        this->functionLibrary->add(ctx->Id()->getText(), std::move(fn));
         return nullptr;
     }
 
@@ -152,13 +162,13 @@ namespace antlrcpptest {
     std::any StrawberryInterpreter::visitArgs(StrawberryParser::ArgsContext *ctx) {
         auto args = std::make_shared<List>();
         for (auto arg: ctx->argument_()) {
-            if (dynamic_cast<StrawberryParser::ArgExpandContext*>(arg) || dynamic_cast<StrawberryParser::RangeArgContext*>(arg)) {
+            if (dynamic_cast<StrawberryParser::ArgExpandContext*>(arg) ||
+                dynamic_cast<StrawberryParser::RangeArgContext*>(arg)) {
                 auto args2 = std::any_cast<std::shared_ptr<Reference>>(visit(arg))->deref()->as<List>();
                 for (int i = 0; i < args2->size(); ++i) {
                     args->append(args2->get(i)->deref());
                 }
-            }
-            else {
+            } else {
                 args->append(std::any_cast<std::shared_ptr<Reference>>(visit(arg))->deref());
             }
         }
@@ -183,7 +193,7 @@ namespace antlrcpptest {
         auto list = std::make_shared<List>();
         int n1 = ref1->toDouble(), n2 = ref2->toDouble();
         if (auto l1 = ref1->deref<List>()) {
-            n1 = l1->get(l1->size()-1)->toDouble();
+            n1 = l1->get(l1->size() - 1)->toDouble();
         }
         if (auto l2 = ref2->deref<List>()) {
             n2 = l2->get(0)->toDouble();
@@ -197,14 +207,14 @@ namespace antlrcpptest {
         return std::make_shared<Reference>(list);
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Regex */ // TODO: fix regex
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Regex */ // TODO: fix regex
     std::any StrawberryInterpreter::visitMatchRegex(StrawberryParser::MatchRegexContext *ctx) {
         return StrawberryParserBaseVisitor::visitMatchRegex(ctx);
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Expressions */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Expressions */
     std::any StrawberryInterpreter::visitLooseFnCallExpr(StrawberryParser::LooseFnCallExprContext *ctx) {
         return StrawberryParserBaseVisitor::visitLooseFnCallExpr(ctx);
     }
@@ -224,7 +234,7 @@ namespace antlrcpptest {
         auto val1 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(0)));
         auto val2 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(1)));
         auto operation = std::any_cast
-                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>>
+                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>>
                 (visit(ctx->op1_()));
         return operation(val1, val2);
     }
@@ -233,7 +243,7 @@ namespace antlrcpptest {
         auto val1 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(0)));
         auto val2 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(1)));
         auto operation = std::any_cast
-                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>>
+                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>>
                 (visit(ctx->op2_()));
         return operation(val1, val2);
     }
@@ -242,7 +252,7 @@ namespace antlrcpptest {
         auto val1 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(0)));
         auto val2 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(1)));
         auto operation = std::any_cast
-                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>>
+                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>>
                 (visit(ctx->op3_()));
         auto val3 = operation(val1, val2);
         return val3;
@@ -252,7 +262,7 @@ namespace antlrcpptest {
         auto val1 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(0)));
         auto val2 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(1)));
         auto operation = std::any_cast
-                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>>
+                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>>
                 (visit(ctx->op4_()));
         return operation(val1, val2);
     }
@@ -261,7 +271,7 @@ namespace antlrcpptest {
         auto val1 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(0)));
         auto val2 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(1)));
         auto operation = std::any_cast
-                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>>
+                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>>
                 (visit(ctx->op5_()));
         return operation(val1, val2);
     }
@@ -270,7 +280,7 @@ namespace antlrcpptest {
         auto val1 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(0)));
         auto val2 = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_(1)));
         auto operation = std::any_cast
-                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>>
+                <std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>>
                 (visit(ctx->op6_()));
         return operation(val1, val2);
     }
@@ -302,8 +312,8 @@ namespace antlrcpptest {
         return visit(ctx->expression_());
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Literals */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Literals */
 
     std::any StrawberryInterpreter::visitDStringLit(StrawberryParser::DStringLitContext *ctx) {
         auto string = ctx->String()->getText();
@@ -376,9 +386,9 @@ namespace antlrcpptest {
         return std::make_shared<Reference>();
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Memory Allocation */
-    // std::shared_ptr<Reference>
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Memory Allocation */
+// std::shared_ptr<Reference>
     std::any StrawberryInterpreter::visitDeclareAssign(StrawberryParser::DeclareAssignContext *ctx) {
         auto varDeclares = ctx->varDeclare_();
         if (varDeclares.size() == 1) {
@@ -432,35 +442,43 @@ namespace antlrcpptest {
     }
 
     std::any StrawberryInterpreter::visitPowAssign(StrawberryParser::PowAssignContext *ctx) {
-        return do_binary_assign_operation(&Value::pow,ctx->identifyer_(),std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
+        return do_binary_assign_operation(&Value::pow, ctx->identifyer_(),
+                                          std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
     }
 
     std::any StrawberryInterpreter::visitMultAssign(StrawberryParser::MultAssignContext *ctx) {
-        return do_binary_assign_operation(&Value::mult,ctx->identifyer_(),std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
+        return do_binary_assign_operation(&Value::mult, ctx->identifyer_(),
+                                          std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
     }
 
     std::any StrawberryInterpreter::visitDivAssign(StrawberryParser::DivAssignContext *ctx) {
-        return do_binary_assign_operation(&Value::div,ctx->identifyer_(),std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
+        return do_binary_assign_operation(&Value::div, ctx->identifyer_(),
+                                          std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
     }
 
     std::any StrawberryInterpreter::visitModAssign(StrawberryParser::ModAssignContext *ctx) {
-        return do_binary_assign_operation(&Value::mod,ctx->identifyer_(),std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
+        return do_binary_assign_operation(&Value::mod, ctx->identifyer_(),
+                                          std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
     }
 
     std::any StrawberryInterpreter::visitAddAssign(StrawberryParser::AddAssignContext *ctx) {
-        return do_binary_assign_operation(&Value::plus,ctx->identifyer_(),std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
+        return do_binary_assign_operation(&Value::plus, ctx->identifyer_(),
+                                          std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
     }
 
     std::any StrawberryInterpreter::visitMinAssign(StrawberryParser::MinAssignContext *ctx) {
-        return do_binary_assign_operation(&Value::min,ctx->identifyer_(),std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
+        return do_binary_assign_operation(&Value::min, ctx->identifyer_(),
+                                          std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_())));
     }
 
     std::any StrawberryInterpreter::visitIncrem(StrawberryParser::IncremContext *ctx) {
-        return do_binary_assign_operation(&Value::plus,ctx->identifyer_(),std::make_shared<Reference>(std::make_shared<Number>(1)));
+        return do_binary_assign_operation(&Value::plus, ctx->identifyer_(),
+                                          std::make_shared<Reference>(std::make_shared<Number>(1)));
     }
 
     std::any StrawberryInterpreter::visitDecrem(StrawberryParser::DecremContext *ctx) {
-        return do_binary_assign_operation(&Value::min,ctx->identifyer_(),std::make_shared<Reference>(std::make_shared<Number>(1)));
+        return do_binary_assign_operation(&Value::min, ctx->identifyer_(),
+                                          std::make_shared<Reference>(std::make_shared<Number>(1)));
     }
 
     std::any StrawberryInterpreter::visitNoInitVarDeclar(StrawberryParser::NoInitVarDeclarContext *ctx) {
@@ -477,15 +495,16 @@ namespace antlrcpptest {
         return declare(ctx->Id()->getText(), ref->deref());
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Identifiers / Container Access */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Identifiers / Container Access */
     std::any StrawberryInterpreter::visitDotAccess(StrawberryParser::DotAccessContext *ctx) {
         auto identifyer = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->identifyer_()));
         if (auto hash = identifyer->get_referenced_value()->as<Hash>()) {
             auto key = std::make_shared<String>(ctx->Id()->getText());
             return hash->get(key);
         }
-        throw std::runtime_error("Cannot use dot access for " + identifyer->deref()->typeName() + "\n\t--->" + ctx->getText());
+        throw std::runtime_error(
+                "Cannot use dot access for " + identifyer->deref()->typeName() + "\n\t--->" + ctx->getText());
     }
 
     std::any StrawberryInterpreter::visitArrAccess(StrawberryParser::ArrAccessContext *ctx) {
@@ -494,7 +513,8 @@ namespace antlrcpptest {
             auto index = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->expression_()));
             return container->get(index->deref());
         }
-        throw std::runtime_error("Cannot access for " + identifyer->deref()->typeName() + " type\n\t--->" + ctx->getText());
+        throw std::runtime_error(
+                "Cannot access for " + identifyer->deref()->typeName() + " type\n\t--->" + ctx->getText());
     }
 
     std::any StrawberryInterpreter::visitArrAccessArgs(StrawberryParser::ArrAccessArgsContext *ctx) {
@@ -503,7 +523,8 @@ namespace antlrcpptest {
             auto index = std::any_cast<std::shared_ptr<Reference>>(visit(ctx->args()));
             return container->get(index->deref());
         }
-        throw std::runtime_error("Cannot access for " + identifyer->deref()->typeName() + " type\n\t--->" + ctx->getText());
+        throw std::runtime_error(
+                "Cannot access for " + identifyer->deref()->typeName() + " type\n\t--->" + ctx->getText());
     }
 
     std::any StrawberryInterpreter::visitIdAccess(StrawberryParser::IdAccessContext *ctx) {
@@ -514,7 +535,7 @@ namespace antlrcpptest {
         return this->get_default();
     }
 
-    // Reference of reference
+// Reference of reference
     std::any StrawberryInterpreter::visitIdReference(StrawberryParser::IdReferenceContext *ctx) {
         return std::make_shared<Reference>(
                 std::any_cast<std::shared_ptr<Reference>>(
@@ -525,14 +546,14 @@ namespace antlrcpptest {
         return StrawberryParserBaseVisitor::visitLooseFnCall(ctx);
     }
 
-    /* -------------------------------------------------------------------------------------------------------------- */
-        /* Operators */
+/* -------------------------------------------------------------------------------------------------------------- */
+/* Operators */
 
-    // -------------------------------------------------------------------------------------------------------------- //
-    // Unary High Priority //
+// -------------------------------------------------------------------------------------------------------------- //
+// Unary High Priority //
     std::any StrawberryInterpreter::visitNegatePrefix(StrawberryParser::NegatePrefixContext *ctx) {
         std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref)->std::shared_ptr<Reference> {
+                operation = [](std::shared_ptr<Reference> ref) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(!ref->toBool()));
         };
         return operation;
@@ -540,7 +561,7 @@ namespace antlrcpptest {
 
     std::any StrawberryInterpreter::visitNegativePrefix(StrawberryParser::NegativePrefixContext *ctx) {
         std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref)->std::shared_ptr<Reference> {
+                operation = [](std::shared_ptr<Reference> ref) -> std::shared_ptr<Reference> {
             auto neg_one = std::make_shared<Number>(-1);
             auto neg_val = ref->mult(neg_one, ref->deref());
             if (neg_val == nullptr) {
@@ -553,7 +574,7 @@ namespace antlrcpptest {
 
     std::any StrawberryInterpreter::visitSizePrefix(StrawberryParser::SizePrefixContext *ctx) {
         std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref)->std::shared_ptr<Reference> {
+                operation = [](std::shared_ptr<Reference> ref) -> std::shared_ptr<Reference> {
             if (auto list = ref->deref<Container>()) {
                 return std::make_shared<Reference>(std::make_shared<Number>(list->size()));
             }
@@ -566,75 +587,83 @@ namespace antlrcpptest {
         return StrawberryParserBaseVisitor::visitExcitedSuff(ctx);
     }
 
-    // -------------------------------------------------------------------------------------------------------------- //
-    // Binary //
-    // std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
+// -------------------------------------------------------------------------------------------------------------- //
+// Binary //
+// std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
     std::any StrawberryInterpreter::visitPowOp(StrawberryParser::PowOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return do_binary_operation(&Value::pow, ref1, ref2);
         };
         return operation;
     }
 
-    // std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
+// std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
     std::any StrawberryInterpreter::visitMultOp(StrawberryParser::MultOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return do_binary_operation(&Value::mult, ref1, ref2);
         };
         return operation;
     }
 
-    // std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
+// std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
     std::any StrawberryInterpreter::visitDivOp(StrawberryParser::DivOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return do_binary_operation(&Value::div, ref1, ref2);
         };
         return operation;
     }
 
-    // std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
+// std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
     std::any StrawberryInterpreter::visitModOp(StrawberryParser::ModOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return do_binary_operation(&Value::mod, ref1, ref2);
         };
         return operation;
     }
 
-    // std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
+// std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
     std::any StrawberryInterpreter::visitPlusOp(StrawberryParser::PlusOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return do_binary_operation(&Value::plus, ref1, ref2);
         };
         return operation;
     }
 
-    // std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
+// std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
     std::any StrawberryInterpreter::visitMinOp(StrawberryParser::MinOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return do_binary_operation(&Value::min, ref1, ref2);
         };
         return operation;
     }
 
     std::any StrawberryInterpreter::visitDefinedOrOp(StrawberryParser::DefinedOrOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             if (!ref1->isNull()) return ref1;
             return ref2;
         };
         return operation;
     }
 
-    // TODO: Redefine bool equals
+// TODO: Redefine bool equals
     std::any StrawberryInterpreter::visitBoolEqOp(StrawberryParser::BoolEqOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             bool bool_result;
             if (ref1->operatorPriority() > ref2->operatorPriority()) {
                 bool_result = ref1->bool_equals(ref1->deref(), ref2->deref());
@@ -647,8 +676,9 @@ namespace antlrcpptest {
     }
 
     std::any StrawberryInterpreter::visitBoolDeepEqOp(StrawberryParser::BoolDeepEqOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             if (ref1->get_referenced_value()->typeName() != ref2->get_referenced_value()->typeName())
                 return std::make_shared<Reference>(std::make_shared<Bool>(false));
             bool bool_result;
@@ -663,8 +693,9 @@ namespace antlrcpptest {
     }
 
     std::any StrawberryInterpreter::visitBoolNeqOp(StrawberryParser::BoolNeqOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             bool bool_result;
             if (ref1->operatorPriority() > ref2->operatorPriority()) {
                 bool_result = !ref1->bool_equals(ref1->deref(), ref2->deref());
@@ -677,8 +708,9 @@ namespace antlrcpptest {
     }
 
     std::any StrawberryInterpreter::visitBoolDeepNeqOp(StrawberryParser::BoolDeepNeqOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             if (ref1->get_referenced_value()->typeName() != ref2->get_referenced_value()->typeName())
                 return std::make_shared<Reference>(std::make_shared<Bool>(false));
             bool bool_result;
@@ -693,40 +725,45 @@ namespace antlrcpptest {
     }
 
     std::any StrawberryInterpreter::visitBoolGtOp(StrawberryParser::BoolGtOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(ref1->toDouble() > ref2->toDouble()));
         };
         return operation;
     }
 
     std::any StrawberryInterpreter::visitBoolGtEqOp(StrawberryParser::BoolGtEqOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(ref1->toDouble() >= ref2->toDouble()));
         };
         return operation;
     }
 
     std::any StrawberryInterpreter::visitBoolLtOp(StrawberryParser::BoolLtOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(ref1->toDouble() < ref2->toDouble()));
         };
         return operation;
     }
 
     std::any StrawberryInterpreter::visitBoolLtEqOp(StrawberryParser::BoolLtEqOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(ref1->toDouble() <= ref2->toDouble()));
         };
         return operation;
     }
 
     std::any StrawberryInterpreter::visitOrOp(StrawberryParser::OrOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(ref1->toBool() || ref2->toBool()));
         };
         return operation;
@@ -737,8 +774,9 @@ namespace antlrcpptest {
     }
 
     std::any StrawberryInterpreter::visitAndOp(StrawberryParser::AndOpContext *ctx) {
-        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>,std::shared_ptr<Reference>)>
-        operation = [](std::shared_ptr<Reference> ref1, std::shared_ptr<Reference> ref2)->std::shared_ptr<Reference> {
+        std::function<std::shared_ptr<Reference>(std::shared_ptr<Reference>, std::shared_ptr<Reference>)>
+                operation = [](std::shared_ptr<Reference> ref1,
+                               std::shared_ptr<Reference> ref2) -> std::shared_ptr<Reference> {
             return std::make_shared<Reference>(std::make_shared<Bool>(ref1->toBool() && ref2->toBool()));
         };
         return operation;
@@ -748,10 +786,10 @@ namespace antlrcpptest {
         return StrawberryParserBaseVisitor::visitBitAndOp(ctx);
     }
 
-    // -------------------------------------------------------------------------------------------------------------- //
-    // Unary Suffix //
+// -------------------------------------------------------------------------------------------------------------- //
+// Unary Suffix //
     std::any StrawberryInterpreter::visitMatchSuff(StrawberryParser::MatchSuffContext *ctx) {
         return StrawberryParserBaseVisitor::visitMatchSuff(ctx);
     }
 
-} // antlrcpptest
+}
