+++ Strawberry/Cpp/main.cpp
@@ -3,8 +3,8 @@
 #include "antlr4-runtime.h"
 #include "StrawberryLexer.h"
 #include "StrawberryParser.h"
-#include "StrawberryInterpreter.h"
 #include "util/Warnings.h"
+#include "StrawberryInterpreter.h"
 
 using namespace antlrcpptest;
 using namespace antlr4;
@@ -16,13 +16,6 @@ int main(const int argc, const char* argv[]) {
     for (int i = 0; i < argc; ++i)
         std::cout << "\t" << i << ": " << argv[i] << std::endl;
     std::cout << std::endl;
-
-    /* DEBUG, if no target is given, default to test.sb */
-    if (argv[1] == nullptr)
-    {
-        argv[1] = R"(C:\Users\Patrick\Documents\Code\ANTLR\StrawberryCpp\Strawberry\test.sb)";
-        std::cout << "Replacing argv[1] with: " << argv[1] << std::endl << std::endl;
-    }
 
     /* Read File and Create Token Stream */
     auto input = ANTLRFileStream();
